MUNASHE NZIRA H210101B TENDEKO BLOCKCHAIN BASED E-PROCUREMENT SYSTEM /backend/app/requirements.txt
fastapi
uvicorn
motor
boto3
pydantic
web3 
eth-account 
python-dotenv
pydantic-settings
PyJWT
passlib
bcrypt
cryptography
sqlalchemy
mysql-connector-python
openai
python-multipart
python-decouple
email-validator